\name{vscc-package}
\alias{vscc-package}
\docType{package}
\title{
Variable selection for clustering and classification
}
\description{
Performs variable selection under a clustering or classification framework. Automated implementation using model-based clustering is based on \code{teigen} version 2.0 and \code{mclust} version 4.0; issues *may* arise when using different versions.
}
\details{
\tabular{ll}{
Package: \tab vscc\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-11-16\cr
License: \tab GPL>=2\cr
}
}
\author{
Jeffrey L. Andrews and Paul D. McNicholas

Maintainer: Jeffrey L. Andrews <jeffrey.andrews@macewan.ca>
}
\references{
See \code{citation("vscc")}.
}
\keyword{ package }
\seealso{
\code{\link{vscc}}
}

