\name{vscc-package}
\alias{vscc-package}
\docType{package}
\title{Variable Selection for Clustering and Classification}
\description{
Performs variable selection under a clustering or classification framework. Automated implementation using model-based clustering is based on \code{teigen} and \code{mclust}.
}
\details{
\tabular{ll}{
Package: \tab vscc\cr
Type: \tab Package\cr
Version: \tab 0.7\cr
Date: \tab 2023-10-17\cr
License: \tab GPL>="2"\cr
}
}
\author{
Jeffrey L. Andrews, Mackenzie Neal, Paul D. McNicholas

Maintainer: Paul D. McNicholas <mcnicholas@math.mcmaster.ca>
}
\references{
See \code{citation("vscc")}.
}
\keyword{package}
\seealso{
\code{\link{vscc}}
}

